refactor(Rust): Create `parser-py` crate
