+++ rust/parser-py/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "parser-py"
+version = "0.0.0"
+edition = "2021"
+
+[dependencies]
+maplit = "1.0.2"
+once_cell = "1.8.0"
+parser-treesitter = { path = "../parser-treesitter", version = "0.0.0" }
+tree-sitter-python = { version = "0.19.0" }
+
+[dev-dependencies]
+test-snaps = { path = "../test-snaps", version = "0.0.0" }
+test-utils = { path = "../test-utils", version = "0.0.0" }
